Removed unnecessary float conversion
x is already a float.

+++ neuralnet/math.go
@@ -3,7 +3,7 @@ package neuralnet
 import "math"
 
 func Sigmoid(x float64) float64 {
-	return 1.0 / (1.0 + math.Pow(math.E, -float64(x)))
+	return 1.0 / (1.0 + math.Pow(math.E, -x))
 }
 
 func SigmoidDerivative(x float64) float64 {
